Pulled out ExchangeArrow in a separate component
