+++ web/src/components/layers/exchangelayer.js
@@ -5,104 +5,15 @@ import React, {
 } from 'react';
 import { useSelector } from 'react-redux';
 import styled from 'styled-components';
-import { noop } from 'lodash';
 
 import { dispatchApplication } from '../../store';
 import { useExchangeArrowsData } from '../../hooks/layers';
 import { useWidthObserver, useHeightObserver } from '../../hooks/viewport';
-import {
-  exchangeQuantizedIntensityScale,
-  exchangeSpeedCategoryScale,
-} from '../../helpers/scales';
 
 import MapExchangeTooltip from '../tooltips/mapexchangetooltip';
+import ExchangeArrow from '../exchangearrow';
 
-// TODO: Fix map scrolling when hovering over arrows when moving map to React.
-// See https://github.com/tmrowco/electricitymap-contrib/issues/2309.
-const ArrowImage = styled.img`
-  cursor: pointer;
-  overflow: hidden;
-  position: absolute;
-  pointer-events: all;
-  image-rendering: crisp-edges;
-  left: -25px;
-  top: -41px;
-`;
-
-const Arrow = React.memo(({
-  arrow,
-  mouseMoveHandler,
-  mouseOutHandler,
-  project,
-  viewportWidth,
-  viewportHeight,
-}) => {
-  const isMobile = useSelector(state => state.application.isMobile);
-  const mapZoom = useSelector(state => state.application.mapViewport.zoom);
-  const colorBlindModeEnabled = useSelector(state => state.application.colorBlindModeEnabled);
-  const {
-    co2intensity,
-    lonlat,
-    netFlow,
-    rotation,
-  } = arrow;
-
-  const imageSource = useMemo(
-    () => {
-      const prefix = colorBlindModeEnabled ? 'colorblind-' : '';
-      const intensity = exchangeQuantizedIntensityScale(co2intensity);
-      const speed = exchangeSpeedCategoryScale(Math.abs(netFlow));
-      return resolvePath(`images/${prefix}arrow-${intensity}-animated-${speed}.gif`);
-    },
-    [colorBlindModeEnabled, co2intensity, netFlow],
-  );
-
-  const transform = useMemo(
-    () => ({
-      x: project(lonlat)[0],
-      y: project(lonlat)[1],
-      k: 0.04 + (mapZoom - 1.5) * 0.1,
-      r: rotation + (netFlow > 0 ? 180 : 0),
-    }),
-    [project, lonlat, rotation, netFlow, mapZoom],
-  );
-
-  const isVisible = useMemo(
-    () => {
-      // Hide arrows with a very low flow...
-      if (Math.abs(netFlow || 0) < 1) return false;
-
-      // ... or the ones that would be rendered outside of viewport ...
-      if (transform.x + 100 * transform.k < 0) return false;
-      if (transform.y + 100 * transform.k < 0) return false;
-      if (transform.x - 100 * transform.k > viewportWidth) return false;
-      if (transform.y - 100 * transform.k > viewportHeight) return false;
-
-      // ... and show all the other ones.
-      return true;
-    },
-    [netFlow, transform],
-  );
-
-  return (
-    <ArrowImage
-      style={{
-        display: isVisible ? '' : 'none',
-        transform: `translateX(${transform.x}px) translateY(${transform.y}px) rotate(${transform.r}deg) scale(${transform.k})`,
-      }}
-      src={imageSource}
-      width="49"
-      height="81"
-      /* Support only click events in mobile mode, otherwise react to mouse hovers */
-      onClick={isMobile ? (e => mouseMoveHandler(arrow, e.clientX, e.clientY)) : noop}
-      onMouseMove={!isMobile ? (e => mouseMoveHandler(arrow, e.clientX, e.clientY)) : noop}
-      onMouseOut={mouseOutHandler}
-      onBlur={mouseOutHandler}
-    />
-  );
-});
-
-const ArrowsContainer = styled.div`
+const Layer = styled.div`
   pointer-events: none;
   position: absolute;
   width: 100%;
@@ -133,7 +44,7 @@ export default React.memo(({ project }) => {
   }, []);
 
   return (
-    <ArrowsContainer id="exchange" ref={ref}>
+    <Layer id="exchange" ref={ref}>
       {tooltip && (
         <MapExchangeTooltip
           exchangeData={tooltip.exchangeData}
@@ -142,8 +53,8 @@ export default React.memo(({ project }) => {
       )}
       {/* Don't render arrows when moving map - see https://github.com/tmrowco/electricitymap-contrib/issues/1590. */}
       {!isMoving && arrows.map(arrow => (
-        <Arrow
-          arrow={arrow}
+        <ExchangeArrow
+          data={arrow}
           key={arrow.sortedCountryCodes}
           mouseMoveHandler={handleArrowMouseMove}
           mouseOutHandler={handleArrowMouseOut}
@@ -152,6 +63,6 @@ export default React.memo(({ project }) => {
           viewportHeight={height}
         />
       ))}
-    </ArrowsContainer>
+    </Layer>
   );
 });
